Move Options, Sources and RuleViolationError into api package.
