+++ djvm/src/main/java/net/corda/djvm/api/package-info.java
@@ -1,4 +1,4 @@
 @Export
-package net.corda.djvm.analysis;
+package net.corda.djvm.api;
 
 import org.osgi.annotation.bundle.Export;
