--- djvm/src/main/java/net/corda/djvm/formatting/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.djvm.formatting;
-
-import org.osgi.annotation.bundle.Export;
