--- djvm/src/main/java/net/corda/djvm/rules/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.djvm.rules;
-
-import org.osgi.annotation.bundle.Export;
