--- djvm/src/main/java/net/corda/djvm/source/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.djvm.source;
-
-import org.osgi.annotation.bundle.Export;
