+++ djvm/src/test/java/net/corda/djvm/execution/AnnotationProxyJavaTest.java
@@ -14,8 +14,8 @@ import net.corda.djvm.Cowboy;
 import net.corda.djvm.TestBase;
 import net.corda.djvm.TypedTaskFactory;
 import net.corda.djvm.WithJava;
+import net.corda.djvm.api.RuleViolationError;
 import net.corda.djvm.rewiring.SandboxClassLoader;
-import net.corda.djvm.rules.RuleViolationError;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
