+++ djvm/src/test/java/net/corda/djvm/execution/ImportTaskJavaTest.java
@@ -1,8 +1,8 @@
 package net.corda.djvm.execution;
 
 import net.corda.djvm.TestBase;
+import net.corda.djvm.api.RuleViolationError;
 import net.corda.djvm.rewiring.SandboxClassLoader;
-import net.corda.djvm.rules.RuleViolationError;
 import org.junit.jupiter.api.Test;
 
 import java.util.function.Function;
