+++ djvm/src/test/java/net/corda/djvm/execution/JavaAccessControllerTest.java
@@ -3,7 +3,7 @@ package net.corda.djvm.execution;
 import net.corda.djvm.TestBase;
 import net.corda.djvm.TypedTaskFactory;
 import net.corda.djvm.WithJava;
-import net.corda.djvm.rules.RuleViolationError;
+import net.corda.djvm.api.RuleViolationError;
 import org.junit.jupiter.api.Test;
 
 import java.security.AccessController;
