+++ djvm/src/test/java/net/corda/djvm/execution/JavaObjectTest.java
@@ -8,8 +8,8 @@ import net.corda.djvm.ExceptionalObjectLongIntConsumer;
 import net.corda.djvm.TestBase;
 import net.corda.djvm.TypedTaskFactory;
 import net.corda.djvm.WithJava;
+import net.corda.djvm.api.RuleViolationError;
 import net.corda.djvm.rewiring.SandboxClassLoader;
-import net.corda.djvm.rules.RuleViolationError;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Method;
