+++ djvm/src/test/java/net/corda/djvm/execution/MaliciousClassLoaderTest.java
@@ -4,8 +4,8 @@ import greymalkin.PureEvil;
 import net.corda.djvm.TestBase;
 import net.corda.djvm.TypedTaskFactory;
 import net.corda.djvm.WithJava;
+import net.corda.djvm.api.RuleViolationError;
 import net.corda.djvm.rewiring.SandboxClassLoader;
-import net.corda.djvm.rules.RuleViolationError;
 import org.junit.jupiter.api.Test;
 
 import java.util.function.Function;
